From List: mapi: remove Xlib/xcb include in gl_marshal.py
The only use of it was to provide the _X_INLINE macro. We already
require (and provide where needed) 'inline', not to mention that it's
used in the file already.

So replace the macro and drop the include. This fixes the build on
platforms which lack the headers - from X-less Linuxes to Androids.

Cc: Timothy Arceri <tarceri@itsqueeze.com>
Cc: Tapani Pälli <tapani.palli@intel.com>
Cc: Vinson Lee <vlee@freedesktop.org>
Reported-by: Vinson Lee <vlee@freedesktop.org>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=100223
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>